Move utils, Xfail on AccuracyValidationException
